# metaclouds-backend-rust/docs/openapi-rust.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Metaclouds Backend API",
    "description": "Rust 版 Metaclouds AI 算力调度后端，对齐 Go v1 API 契约。\n\n统一响应信封：`{\"success\":bool,\"data\":...,\"message\":...,\"code\":...,\"timestamp\":sec}`。错误响应对见 `ErrorResponse`。",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "development"
    }
  ],
  "paths": {
...
    "/api/v1/quotas": {
      "get": {
        "tags": [
          "quotas"
        ],
        "summary": "`GET /api/v1/quotas` — 分页列表（过滤）。",
        "operationId": "list_quotas",
        "responses": {
          "200": {
            "description": "paginated quotas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaPage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "quotas"
        ],
        "summary": "`POST /api/v1/quotas` — 创建（201）。",
        "operationId": "create_quota",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuotaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceQuotaResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/quotas/{id}": {
      "get": {
        "tags": [
          "quotas"
        ],
        "summary": "`GET /api/v1/quotas/:id` — 详情。",
        "operationId": "get_quota",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "quota",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceQuotaResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "quotas"
        ],
        "summary": "`PUT /api/v1/quotas/:id` — 更新。",
        "operationId": "update_quota",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuotaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceQuotaResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "quotas"
        ],
        "summary": "`DELETE /api/v1/quotas/:id` — 软删除，204。",
        "operationId": "delete_quota",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/quotas/{id}/check": {
      "post": {
        "tags": [
          "quotas"
        ],
        "summary": "`POST /api/v1/quotas/:id/check` — 检查是否超限。",
        "operationId": "check_quota",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckQuotaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "check result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaCheckResult"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckQuotaRequest": {
        "type": "object",
        "description": "`POST /api/v1/quotas/:id/check` 请求体（资源请求量）。",
        "properties": {
          "cpu": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "gpu": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "memory_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "storage_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
      "CreateQuotaRequest": {
        "type": "object",
        "description": "`POST /api/v1/quotas` 请求体。",
        "required": [
          "name",
          "tenant_id"
        ],
        "properties": {
          "cpu_limit": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "gpu_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "memory_limit_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "partition_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "storage_limit_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "统一错误响应 schema（对齐 `ApiResponse` 错误形式）。\n\n实际信封：`{\"success\":false,\"message\":\"...\",\"code\":\"...\",\"timestamp\":<sec>}`。\n成功响应的 `data` 字段由各 handler 对应的业务 schema 描述。",
        "required": [
          "success",
          "message",
          "code",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "QuotaCheckResult": {
        "type": "object",
        "description": "check_quota 结果：是否允许 + 超限项列表。",
        "required": [
          "allowed",
          "exceeded"
        ],
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "exceeded": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "QuotaPage": {
        "type": "object",
        "description": "分页配额列表响应内层。",
        "required": [
          "data",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceQuotaResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "page_size": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ResourceQuotaResponse": {
        "type": "object",
        "description": "对外配额视图（剔除软删除列 `deleted_at`）。",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "description",
          "tenant_id",
          "gpu_limit",
          "gpu_used",
          "cpu_limit",
          "cpu_used",
          "memory_limit_gb",
          "memory_used_gb",
          "storage_limit_gb",
          "storage_used_gb",
          "status"
        ],
        "properties": {
          "cpu_limit": {
            "type": "number",
            "format": "double"
          },
          "cpu_used": {
            "type": "number",
            "format": "double"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "gpu_limit": {
            "type": "integer",
            "format": "int64"
          },
          "gpu_used": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "memory_limit_gb": {
            "type": "number",
            "format": "double"
          },
          "memory_used_gb": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "partition_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "storage_limit_gb": {
            "type": "number",
            "format": "double"
          },
          "storage_used_gb": {
            "type": "number",
            "format": "double"
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateQuotaRequest": {
        "type": "object",
        "description": "`PUT /api/v1/quotas/:id` 请求体（全部可选）。",
        "properties": {
          "cpu_limit": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "gpu_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "memory_limit_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "storage_limit_gb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
}